Refactor and add redirects to partitions
