+++ 07_distributed_systems/01_intro/server/wal.go
@@ -4,6 +4,7 @@ import (
 	"encoding/gob"
 	"fmt"
 	"io"
+	"kv_store/helpers"
 	"log"
 	"net/http"
 	"os"
@@ -23,7 +24,7 @@ func addToWAL(server *serverState, key, value string, w http.ResponseWriter) err
 		Value: value,
 	}
 	err := server.walGobEncoder.Encode(&wr)
-	if errorResponse(&w, err, http.StatusInternalServerError) != nil {
+	if helpers.ErrorResponse(&w, err, http.StatusInternalServerError) != nil {
 		server.mutex.Unlock()
 		return err
 	}
@@ -64,18 +65,18 @@ func readWAL(server *serverState, minId uint64) (inMemoryStorage, error) {
 func (server *serverState) walGetHandler(w http.ResponseWriter, r *http.Request) {
 	walNumbers, ok := r.URL.Query()["since"]
 	if !ok {
-		errorResponse(&w,
+		helpers.ErrorResponse(&w,
 			fmt.Errorf("since parameter is not found in the request"),
 			http.StatusBadRequest)
 		return
 	}
 
 	walNumber, err := strconv.ParseUint(walNumbers[0], 10, 64)
-	if errorResponse(&w, err, http.StatusBadRequest) != nil {
+	if helpers.ErrorResponse(&w, err, http.StatusBadRequest) != nil {
 		return
 	}
 	diff, err := readWAL(server, walNumber)
-	if errorResponse(&w, err, http.StatusInternalServerError) != nil {
+	if helpers.ErrorResponse(&w, err, http.StatusInternalServerError) != nil {
 		return
 	}
 	encoder := gob.NewEncoder(w)
